requirement.txt
django==3.1.1
djangorestframework==3.12.1
pillow==8.0.1
django-filter==2.4.0